requirements.txt
ipdb
ipykernel
matplotlib
seaborn
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.1
google-cloud-storage
geopandas
pathlib
msgpack
s2cell
s2sphere
msgpack
torch
torchvision
opencv-python
gcsfs
fsspec
fastapi
folium
st_clickable_images
streamlit_folium
streamlit_folium
